Only allow valid MRNs
MRNs have a specific format.  They consist of:
* the two-digit year (e.g. 19)
* the two-character country code (e.g. GB)
* 13 further characters which can be digits or uppercase chars
* A check digit calculated per ISO 6346

This PR changes the MovementReferenceNumber model to have an
apply method which will validate both the format and the check
digit, and changes both the Json Reads and the path bindable
to use it and fail if given an invalid MRN.  It also changes
the MovementReferenceNumber form to validate.