CTCTRADERS-144 Remove index from Add Event page url and use answers.
This removs the index from the index of the url on the Add Event Page
and instead depending on the userAnswers to determine what the index
should be. To calculate the index we use NumberOfEvents to calculate
the number of events.

This also uses NumberOfEvents to determine how many rows to generate
for the add event list page.