Refactor Section title to be Text vs String.
This change removes a common requirement of having messages in scope
when constructing the view model. Instead this message lookup and
translation is differed to serialization to json, typically in the
controllers where request is available and messages can be derived.

This is part of a design decision to move away from passing implicit
requests or messages around and polluting function signatures.